Implement hack to _remove_ binaries from pyinstaller analysis.
By removing most of the shared binaries the resulting executable should be
more portable and flexible than the last using the builtin system libraries.
Tested by building on Ubuntu 16 and works on both 18 and 20.

Should work as long as GTK / Webkit2 are available.

This also has the added advantage of generating much smaller (>50%) and
quicker Linux executables.